+++ js/controls/data-table.js
@@ -189,7 +189,7 @@
                         throw Error('Error - When using <data-table> with a template the script [js/extensions/jsTemplate.js] is required.');
                     }
                     // Render each item using the template
-                    var render = app.jsTemplate.compile(null, template.innerHTML);
+                    var render = app.jsTemplate.compile(null, false, template.innerHTML);
                     for (var index = 0, count = list.length; index < count; index++) {
                         item = list[index];
                         try {