+++ js/web-components/url-router.js
@@ -47,7 +47,7 @@ class UrlRouter extends HTMLElement {
         }
         const shadowRoot = this.attachShadow({mode: 'open'});
         shadowRoot.appendChild(shadowTmpl.content.cloneNode(true));
-        this._currentRoute = null;
+        this.currentRoute = null;
         this.updateView = this.updateView.bind(this);
         this.updateView();
     }
@@ -60,10 +60,6 @@ class UrlRouter extends HTMLElement {
         window.removeEventListener('popstate', this.updateView);
     }
 
-    get currentRoute() {
-        return this._currentRoute;
-    }
-
     /**
      * Called when the element is added to the page
      * and when the URL #hash changes.
@@ -74,7 +70,7 @@ class UrlRouter extends HTMLElement {
         await componentsAreDefined(this, 'url-route');
 
         // Setup Routing
-        this._currentRoute = null;
+        this.currentRoute = null;
         const view = setupRouting(this);
 
         // Get URL Path
@@ -109,7 +105,7 @@ class UrlRouter extends HTMLElement {
                 // the HTML will be downloaded and set to <template> the
                 // first time the route is accessed.
                 const html = route.template;
-                this._currentRoute = route;
+                this.currentRoute = route;
                 if (html === null) {
                     downloadTemplate(this, view, route, result.urlParams);
                 } else {